Day 2 part 1 completion
Removing extra text from day1 that clutters up tests now. I didn't come up with useful text here but one or 2 streams of messages doesn't really work here either. Figured out how to assign count as an enumerable of the for loop and to use Enum.sum then. To my knowledge, the thing with for loops is unlike something like PHP, the inner section does not allow us to capture or accumulate at this scope. I'm sure there's an easier way but my google fu missed it. The `character_count()` `Enum.reduce` usage works up an accumulator via recursion that is roughly what we're doing. These patterns are becoming easier to focus on but I still don't live in them as well as other languages.