Day 2 part 2 completion
Ugly duplication since I had no clue the changed part would be nested deep but ideally I think I can start to see how to structure these or refactor them later. This isn't meant to be pretty and this is taking damn near forever because Elixir is still foreign.

Verified the answer is complete and added part 2 to the puzzle. It would be useful to have versions without answers or with them redacted? Some way to reset and redo these like elixir koans.