Day 3 part 1 completion (final)
Verified this produced the right answer. Had no clue I needed to convert the Stream Enum to a list, or at least for this example. What I assumed was always a list was mostly just enumerable in a similar form. Unfortunately I took 2 hints for this, how to extend the map and the use of modulo. I was seconds away from each implementation but my hang up stems from the puzzle text. This is always enough information but my brain doesn't always see the way through each maze immediately but with a few breadcrumbs, I work out each portion. There's honestly a lot of and enough work just getting Elixir to do my bidding and this is ugly enough to be completely mine.