Day 4 part 1 parsing passports
Worked out Stream transform example we needed further up the chain to chunk files in and split on 2 new lines. This isn't the most perfect example since regex works a little cleaner considering the different environments use `\r`, `\n` or `\r\n`. Now that we have our passports as a list of strings, we can do the field parsing required for the rest.