Day 4 part 1 completion
Modified the input to end with 2 new lines, otherwise the last passport is always skipped via this method. I can't tell if this is against the rules but this is what I would do personally more often than not. Commented out all of our debugging lines. Using `key in valid_keys` syntax to build a list of booleans that Enum.all verifies are all true before kicking it upstream to count only the valid passports.

I feel like I'm kind of producing side effects or I just don't know what that phrase means. I'm getting the final answer but I can't tell if this is a good usage pattern or not. The good news of today is this is all me, no hints other than Google figuring out how to chunk in Elixir or random topics to help me push through it.