Day 4 part 1 incorrect
Figured out how to replace in the Stream usage to fix one issue upstream. Turns out I had been doing this all wrong, my solution was looking for keys in the list. We needed to do the opposite, look for and get a list of keys that are not present in the passport. We're at an intermediate step where we have a list of MapSet and we need to focus on counting the ones that are empty.