Day 4 part 2 completion
Add the right answer after a little trial and error. Had to look at examples on https://elixirforum.com/t/advent-of-code-2020-day-4/35971/4 for the final hint. The in syntax is cleaner to look at and I really wasn't doing comparisons correctly because every value was a string when I did integer checking. This is happening quite a lot tbh and its one of the smaller yet annoying things to deal with. PHP and other languages kind of coerce to making this work when I really just need to be more explicit.