Day 5 part 2 completion
Attempted to work through solution from https://github.com/Maximization/adventofcode2020/blob/main/day5/index.js but had issues trying to make Enum.reduce work. I came up with the current solution solo and tbh I'm surprised it worked. This builds a second list of all rows from the min to the max seat, performs the same MapSet difference and returns the only missing value. This portion of the puzzle is what helped it stand out though: `However, there's a catch: some of the seats at the very front and back of the plane don't exist on this aircraft, so they'll be missing from your list as well.`. Because our min is 46 and not 0, we don't have the seats that are missing.