Day 6 part 1 input parsing
Splitting groups based on the same parsing from day 4. This splits the total grouping as one big list with individual lists for the remaining new lines so our test data so far is `[["abc"], ["a", "b", "c"], ["ab", "ac"], ["a", "a", "a", "a"], ["b"]`. We again fudge the data since the last line isn't a double new line, even though we always end in a new line for git reasons.