Day 6 part 1 input grouping
Now we group even further looking for String.graphemes to build even smaller lists. This builds a result set of:
```elixir
Groups: [
  [["a", "b", "c"]],
  [["a"], ["b"], ["c"]],
  [["a", "b"], ["a", "c"]],
  [["a"], ["a"], ["a"], ["a"]],
  [["b"]]
]
```
From here we can look at each inner grouping and combine graphemes and lists with uniqueness to make sure the first shows a+b+c (3), the second and the third should be the same. The unique constraint should strip the 4th group of a's to 1 and the last to 1 as well, giving us the 11 we're looking for. I think it's also possible to work with just the Answers set by combining what look like graphemes into strings and reducing any repeated characters.