Day 6 part 2 test complete
Realized this is building on the earlier implementation from day 2 to count the number of characters in a string. I get the count of the enum to determine how many people should be present and inspect each character to see if the count matches. If it does, we add to the accumulator. The else statement here keeps us at zero instead of nil that it would default to.