Day 7 part 1 implementation halt
Building a list of rules but from the hints I've looked for, this is a breadth first search vs a depth first search. One approach I was looking at was to loop over this result to combine all the counts and do a parent -> child -> child exponential count but I think that's wrong now because counts of parents will be different. Like if one bag has 4 and another has 3, making it 7 is naive and wrong depending on where in the graph it is. I'm going to jump to the next day and likely jump back to this with cheats. I want to keep the structure I built and even merging someone else's into that data structure is enough of a learning experience.