Day 8 part 1 complete
Unpacked implementation at https://github.com/rkachowski/advent-of-code/blob/master/2020/8/solution.exs using case statement instead of function declarations. This is a bit of a cheat but I did work out that you needed a list of seen indexes as you jump around. Once you've been seen, you're back at an infinite loop. This example reinforces that Elixir types are immutable but to accumulate an object or result, passing it as an argument is a useful tactic. Impostor syndrome hit pretty hard for this tbh as my brain can't parse into concise functional concepts like this.