Day 8 part 2 complete (mega cheat)
Pretty much lifted the result from https://github.com/rkachowski/advent-of-code/blob/master/2020/8/solution.exs though I need to explain precisely what it's doing if I expect to actually learn instead of just cheat my way to solutions. It looks like the `Enum.reduce` function parses the instructions down to the indexes and when it's either a jmp or nop instruction. The `List.update_at` switches values based on the pattern matched and it proceeds to do the check from part 1 but looking for the end because as it detects a loop it recurses back through the function with the next index in the tail. Currently, brute_force runs once with the test data but it looks to loop with the input. I may just end up working through these even if I borrow heavily or use other people's work. Part of it is because I don't want to devote too much beyond 30 days to this initially. My plan is to refactor everything and start over after I've skilled up because I feel on the n00b side. You can do a lot on the web and never reach problems anywhere near this magnitude.