Day 9 part 1 incorrect
I wanted to commit the journal and incorrect code as I hadn't read the problem correctly and it was only exposed by looking at the output. What is supposed to happen is a sliding window of sorts so for the first check it's the entire preamble, the second it slices the first off the preamble and covers the last total, continuing until the last number.

There's a recursive function here, something that likely doesn't use Enum.split because I don't think you can give it an offset. Split works for our first check and is useful to capture only the portion we iterate over but we need the whole input to split as we iterate. I think I'm finally going to take a hint but do my best to keep it a clean room implementation or just take inspiration in how the iteration happens.