Day 9 part 1 complete
While there's little that seems attributed to me from this implementation of https://github.com/LostKobrakai/aoc2020/blob/master/lib/aoc2020/day9.ex#L6, it is the closest to what I had worked up so far. The use of `Stream.chunk_every` is what I was missing to build a "sliding window" and the logic in `Enum.find_every` is very close to my implementation. I honestly don't know if I would've worked this out but all of the other examples (more in the journal entry) are vastly different than what I started with.